PLAYG-64 Notice 안으로 post, comment 이동
