+++ src/main/java/com/playground/api/notice/controller/PostController.java
@@ -1,4 +1,5 @@
-package com.playground.api.post.controller;
+package com.playground.api.notice.controller;
+
 
 import java.util.List;
 
@@ -10,9 +11,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.playground.api.post.model.PostRequest;
-import com.playground.api.post.model.PostResponse;
-import com.playground.api.post.service.PostService;
+import com.playground.api.notice.model.PostRequest;
+import com.playground.api.notice.model.PostResponse;
+import com.playground.api.notice.service.PostService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
