+++ src/main/java/com/playground/api/notice/entity/PostEntity.java
@@ -1,4 +1,4 @@
-package com.playground.api.comment.entity;
+package com.playground.api.notice.entity;
 
 import com.playground.entity.BaseEntity;
 
@@ -22,22 +22,18 @@ import lombok.RequiredArgsConstructor;
 @EqualsAndHashCode(callSuper = true)
 @Getter
 @Entity
-@Table(name = "tb_comment")
-@IdClass(CommentEntityPK.class)
+@Table(name = "tb_post")
+@IdClass(PostEntityPK.class)
 @SequenceGenerator(
-	    name = "cmnt_no_seq",
-	      sequenceName = "tb_comment_cmnt_no_seq",
+	    name = "ntt_no_seq",
+	      sequenceName = "tb_post_ntt_no_seq",
 	      initialValue = 1,
 	      allocationSize = 1
 	  )
-public class CommentEntity extends BaseEntity{
-	
+public class PostEntity extends BaseEntity{
+
 	@Id
-	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator="cmnt_no_seq")
-	@Column(name = "cmnt_no")
-	private int cmntNo;
-	
-	@Id
+	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator="ntt_no_seq")
 	@Column(name = "ntt_no")
 	private int nttNo;
 	
@@ -45,11 +41,18 @@ public class CommentEntity extends BaseEntity{
 	@Column(name = "bbs_id")
 	private String bbsId;
 	
-	@Column(name = "cmnt_cn")
-	private String cmntCn;
+	@Column(name = "ntt_sj")
+	private String nttSj;
 	
-	@Column(name = "upper_cmnt_no")
-	private int upperCmntNo;
+	@Column(name = "ntt_cn")
+	private String nttCn;
+	
+
+	@Column(name = "regist_usr_id")
+	private String registUsrId;
+	
+	@Column(name = "updt_usr_id")
+	private String updtUsrId;
 	
 	
 }
