+++ src/main/java/com/playground/api/notice/entity/PostEntityPK.java
@@ -1,4 +1,4 @@
-package com.playground.api.post.entity;
+package com.playground.api.notice.entity;
 
 import java.io.Serializable;
